refactor: test내부 클래스 stub 외부로 구조 변경
